Update to Asterisk SVN trunk r24950
------------------------------------------------------------------------
r24912 | russell | 2006-05-05 19:39:31 +0200 (Fri, 05 May 2006) | 1 line


------------------------------------------------------------------------
r24950 | russell | 2006-05-05 20:11:55 +0200 (Fri, 05 May 2006) | 10 lines

re-add the initialization of the scheduled item's time to 0.  I had removed
this because I checked the sched_alloc function to use calloc instead of
malloc, so I thought it was no longer necessary.  However, the sched structures
are cached, and cached values will have the old values in them, so this still
needs to be done.

Also, wrap the scheduler debug code to only happen if option_debug is enabled.
It spits out a ton of output so it's nice to be able to enable/disable it
during runtime.

------------------------------------------------------------------------