Update to Asterisk SVN trunk r32763
------------------------------------------------------------------------
r32759 | russell | 2006-06-07 01:22:12 +0200 (Wed, 07 Jun 2006) | 7 lines

revert my changes to convert fields of zt_pvt to use stringfields.  When I did
this, I was not keeping in mind the fact that after a stringfield is overwritten
by another string, the memory used by the old string can not be recovered.  I
would like to go back through these changes and make sure that stringfields are
not used for fields that are written to many times before these changes are
committed.

------------------------------------------------------------------------