Update to Asterisk SVN trunk r35443
------------------------------------------------------------------------
r35436 | russell | 2006-06-22 14:14:29 +0200 (Thu, 22 Jun 2006) | 9 lines

instead of using a for loop to call make on each subdirectory, use a target for
each subdirectory.
 - the shell code in the for loop doesn't have a manually catch errors from
   make on a subdirectory
 - you can now run something like "make apps" or "make channels" from the root
   source directory
 - This should allow for better support for parallel instances of make (the -j
   option) since make can work on each subdirectory target in parallel

------------------------------------------------------------------------
r35437 | russell | 2006-06-22 14:33:50 +0200 (Thu, 22 Jun 2006) | 3 lines

instead of using a for loop to call 'make install' on each subdirectory, use
separate targets

------------------------------------------------------------------------
r35438 | russell | 2006-06-22 14:48:25 +0200 (Thu, 22 Jun 2006) | 3 lines

use a slightly different, but equivalent way of setting the 
SUBDIRS_INSTALL variable

------------------------------------------------------------------------
r35439 | russell | 2006-06-22 14:55:49 +0200 (Thu, 22 Jun 2006) | 5 lines

revert my change to use the AC_PREFIX_DEFAULT macro since it is not something
that is expanded inline, so it didn't do what I wanted it to. The --help output
will be wrong again, but that's better than completely breaking it. :)
(Thanks to jcollie for catching this!)

------------------------------------------------------------------------